requirements.txt
kafka-python >= 1.4.3
ubirch-python-utils >= 1.0.6